requirements.txt
coreapi==2.3.3
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
django-environ==0.9.0
